\docType{package}
\name{FAOProductionImputation-package}
\alias{FAOProductionImputation-package}
\title{Package to perform the imputation of area, production and
yield for the FAO production domain.}
\description{
  Package to perform the imputation of area, production and
  yield for the FAO production domain.
}
\author{
  Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

